implement the processor further and modulize the plugin
now it syncs new tasks from and to vikunja and vault. But there are no removal mechanic. Also there are bugs with ID handling.
Also it updates tasks from vault to vikunja, but not in reverse, because there is a need for a time mechanic, how to track the updated time.