now labels can be added to tasks and will be removed from them. Also deleting tasks and labels automatically, if they are not mentioned in the vault.
this adds new settings for automations. Also it adds a button in settings to trigger check for dependencies as it is triggered on plugin loading.
Also it implements update process for tasks in vikunja from vault changes. But not in reverse.
Also it implements the correct check for due date and done date.