cache is only available for tasks, which have an id
so the processor runs over the whole vault and uses modified date for creating, but after updates, will be used the cache and helps parser to take the correct updated value.

the cache looks always at the whole file and add them to cache. also the pulling stuff will be cached.